+++ Koala/koala_lang/src/lang/Parser.cpp
@@ -61,7 +61,7 @@ namespace KoalaLang{
                 if(PARSER_CURRENT_TOKEN.value == "ret"){
                     Next();
                     //TODO: parsing expressions
-                    SHARED_PTR_T(ASTNumberLiteral) constant = std::make_shared<ASTNumberLiteral>(std::stod(PARSER_CURRENT_TOKEN.value));
+                    SHARED_PTR_T(ASTNumberLiteral) constant = std::make_shared<ASTNumberLiteral>(std::stoul(PARSER_CURRENT_TOKEN.value));
                     FatalNext(TokenType::Semicolon);
                     block->GetNodes().push_back(std::make_shared<ASTRet>(constant));
                 }
